Add 3.x test
