+++ impl/maven-cli/src/main/java/org/apache/maven/cling/invoker/mvn/embedded/EmbeddedMavenExecutor.java
@@ -118,6 +118,7 @@ public class EmbeddedMavenExecutor implements Executor {
             properties.put(
                     "user.dir",
                     executorRequest.cwd().toAbsolutePath().normalize().toString());
+            properties.put("maven.multiModuleProjectDirectory", executorRequest.cwd().toAbsolutePath().normalize().toString());
             properties.put(
                     "user.home",
                     executorRequest
