+++ impl/maven-cli/src/test/java/org/apache/maven/cling/invoker/mvn/MavenInvokerTestSupport.java
@@ -44,57 +44,11 @@ public abstract class MavenInvokerTestSupport {
                 "${maven.home}/conf/maven.properties must be a file");
 
         Files.createDirectory(cwd.resolve(".mvn"));
-
-        String pomString =
-                """
-                    <?xml version="1.0" encoding="UTF-8"?>
-                    <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-                             xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/maven-v4_0_0.xsd">
-
-                        <modelVersion>4.0.0</modelVersion>
-
-                        <groupId>org.apache.maven.samples</groupId>
-                        <artifactId>sample</artifactId>
-                        <version>1.0.0</version>
-
-                        <dependencyManagement>
-                          <dependencies>
-                            <dependency>
-                              <groupId>org.junit</groupId>
-                              <artifactId>junit-bom</artifactId>
-                              <version>5.11.1</version>
-                              <type>pom</type>
-                              <scope>import</scope>
-                            </dependency>
-                          </dependencies>
-                        </dependencyManagement>
-
-                        <dependencies>
-                          <dependency>
-                            <groupId>org.junit.jupiter</groupId>
-                            <artifactId>junit-jupiter-api</artifactId>
-                            <scope>test</scope>
-                          </dependency>
-                        </dependencies>
-
-                    </project>
-                    """;
         Path pom = cwd.resolve("pom.xml").toAbsolutePath();
-        Files.writeString(pom, pomString);
-
-        String appJavaString =
-                """
-                package org.apache.maven.samples.sample;
-
-                public class App {
-                    public static void main(String... args) {
-                        System.out.println("Hello World!");
-                    }
-                }
-                """;
+        Files.writeString(pom, MavenTestSupport.POM_STRING);
         Path appJava = cwd.resolve("src/main/java/org/apache/maven/samples/sample/App.java");
         Files.createDirectories(appJava.getParent());
-        Files.writeString(appJava, appJavaString);
+        Files.writeString(appJava, MavenTestSupport.APP_JAVA_STRING);
 
         Parser parser = createParser();
         try (Invoker invoker = createInvoker()) {
