+++ impl/maven-cli/src/test/java/org/apache/maven/cling/invoker/mvn/MavenTestSupport.java
@@ -1,0 +1,69 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.apache.maven.cling.invoker.mvn;
+
+public final class MavenTestSupport {
+    private MavenTestSupport() {}
+
+    public static final String POM_STRING =
+            """
+                <?xml version="1.0" encoding="UTF-8"?>
+                <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+                         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/maven-v4_0_0.xsd">
+
+                    <modelVersion>4.0.0</modelVersion>
+
+                    <groupId>org.apache.maven.samples</groupId>
+                    <artifactId>sample</artifactId>
+                    <version>1.0.0</version>
+
+                    <dependencyManagement>
+                      <dependencies>
+                        <dependency>
+                          <groupId>org.junit</groupId>
+                          <artifactId>junit-bom</artifactId>
+                          <version>5.11.1</version>
+                          <type>pom</type>
+                          <scope>import</scope>
+                        </dependency>
+                      </dependencies>
+                    </dependencyManagement>
+
+                    <dependencies>
+                      <dependency>
+                        <groupId>org.junit.jupiter</groupId>
+                        <artifactId>junit-jupiter-api</artifactId>
+                        <scope>test</scope>
+                      </dependency>
+                    </dependencies>
+
+                </project>
+                """;
+
+    public static final String APP_JAVA_STRING =
+            """
+            package org.apache.maven.samples.sample;
+
+            public class App {
+                public static void main(String... args) {
+                    System.out.println("Hello World!");
+                }
+            }
+            """;
+}
